Connected SignUpPage to Database
