+++ easyPdf/db/bzl/userBzl.py
@@ -33,7 +33,14 @@ def create_user(request):
         user_dto = UserDto(data=request.data)
         print(user_dto)
         if user_dto.is_valid():
-            user_dto.save()
+            first_name = user_dto.data.get('first_name')
+            last_name = user_dto.data.get('last_name')
+            username = user_dto.data.get('username')
+            password = user_dto.data.get('password')
+            email = user_dto.data.get('email')
+            user = User(first_name=first_name, last_name=last_name, username=username, email=email)
+            user.set_password(password)
+            user.save()
             return Response("User has been created!", status=status.HTTP_200_OK)
         return Response('User DTO not VALID OR ALREADY EXISTS(e.g. admin mindfuck)', status=status.HTTP_409_CONFLICT)
     return Response('BAD REQUEST', status=status.HTTP_400_BAD_REQUEST)
