+++ easyPdf/frontend/bzl/signup/SignupBzl.js
@@ -2,8 +2,8 @@ import React from 'react';
 import {UserAPI} from "../../api/user/UserApi";
 
 
-export const logUserIn = async (user) => {
-    const response = await UserAPI.logUserIn(user);
+export const signUserUp = async (user) => {
+    const response = await UserAPI.signupUser(user);
     console.log(`Response value: [${response.ok}]`)
     return response.ok;
 }
